refactor(ledger): collocate the operation-route route registrars :hammer:
The Huma terminal registrar came out of operation_route_handler.go and the
Fiber guard-chain registrars came out of routes.go; both now live in
operation_route_routes.go.

No route, authz tuple, operation ID or path changed.
