+++ components/ledger/internal/adapters/http/in/operation_route_handler.go
@@ -243,75 +243,3 @@ func (handler *OperationRouteHandler) DeleteOperationRouteByID(ctx context.Conte
 
 	return &DeleteOperationRouteResponse{}, nil
 }
-
-// RegisterOperationRouteRoutes registers the five operation-route operations on the
-// shared Huma API. It is the per-file seam the unified server calls; the auth
-// ("midaz","operation-routes",verb) + tenant +
-// ParseUUIDPathParameters("operation_route") middleware chain is attached on the
-// version group (Fiber-level) BEFORE the Huma terminal, not here. Paths are
-// GROUP-RELATIVE (see asset_handler.go's RegisterAssetRoutes header for the
-// rationale).
-//
-// opSuffix is appended to every operation ID so the same surface can be published on
-// more than one version group of the one document without colliding — see
-// routeOpSuffixV1.
-func RegisterOperationRouteRoutes(api huma.API, h *OperationRouteHandler, opSuffix string) {
-	const (
-		listPath = "/organizations/{organization_id}/ledgers/{ledger_id}/operation-routes"
-		idPath   = listPath + "/{operation_route_id}"
-		tag      = "Operation Routes"
-	)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "createOperationRoute" + opSuffix,
-		Method:      http.MethodPost,
-		Path:        listPath,
-		Summary:     "Create Operation Route",
-		Tags:        []string{tag},
-		Security:    secOperationRouteBearer,
-		// Body validated imperatively (http.DecodeAndValidate) — see file header.
-		SkipValidateBody: true,
-		DefaultStatus:    http.StatusCreated,
-	}, h.CreateOperationRoute)
-	attachTypedRequestBody[mmodel.CreateOperationRouteInput](api, "createOperationRoute"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "listOperationRoutes" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        listPath,
-		Summary:     "Retrieve all operation routes",
-		Tags:        []string{tag},
-		Security:    secOperationRouteBearer,
-	}, h.GetAllOperationRoutes)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getOperationRouteByID" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        idPath,
-		Summary:     "Retrieve a specific operation route",
-		Tags:        []string{tag},
-		Security:    secOperationRouteBearer,
-	}, h.GetOperationRouteByID)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "updateOperationRoute" + opSuffix,
-		Method:           http.MethodPatch,
-		Path:             idPath,
-		Summary:          "Update an operation route",
-		Tags:             []string{tag},
-		Security:         secOperationRouteBearer,
-		SkipValidateBody: true, // body validated imperatively — RFC 7396 merge-patch core.
-	}, h.UpdateOperationRoute)
-	attachTypedRequestBody[mmodel.UpdateOperationRouteInput](api, "updateOperationRoute"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "deleteOperationRoute" + opSuffix,
-		Method:      http.MethodDelete,
-		Path:        idPath,
-		Summary:     "Delete an operation route",
-		Tags:        []string{tag},
-		Security:    secOperationRouteBearer,
-		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
-		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteOperationRouteByID)
-}
